==> 04c_Distributions_alfa_beta_q.ipynb <==
# Import libraries

import os

import pandas as pd
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
from scipy.stats import norm, truncnorm, multivariate_normal
from sklearn.linear_model import LinearRegression
import corner
import pylab
from matplotlib import colors

%matplotlib inline

import pickle

# Functions

def load_dict_from_file(name):
    a_file = open(name, "rb")
    output = pickle.load(a_file)
    a_file.close()
    return output

style_param = {'legend.fontsize': '20',
         'xtick.direction' : 'in',  
         'ytick.direction' : 'in', 
         'xtick.top' : True, 
         'figure.figsize': (10,6),
         'axes.labelsize': '26',
         'axes.titlesize':'26',
         'xtick.labelsize':'20',
         'ytick.labelsize':'20',
         'xtick.major.pad':'16',
         'ytick.major.pad':'16'}
    
pylab.rcParams.update(style_param)

# Open data 

d = load_dict_from_file('tmp/sample1_dict_alpha_beta.pkl')

# Alpha & Beta distributions

#choose the experiment to visualize the results from article 10, 
#where a proton energy of 5010keV was used
from matplotlib.colors import LogNorm


a = d['10'][5010]['alpha_dep']
b=(d['10'][5010]['beta_dep'])

fig, ax = plt.subplots(2,2, figsize=(8,8))

# plot 1D histograms
ax[0][0].hist(a, bins=100, density=True,color='blue',alpha=0.4)
ax[0][0].set_xlim(0.2,0.4)

ax[1][1].hist(b, bins=100, density=True, orientation='horizontal',color='blue',alpha=0.4)
ax[1][1].set_ylim(0.,0.05)


ax[0][0].set_title(r'$\alpha\; [Gy^{-1}]$')
ax[0][0].xaxis.set_ticks_position('top')
ax[0][0].set_ylabel('Probability\n distribution[-]\n')

ax[1][1].set_ylabel('\n'+r'$\beta\; [Gy^{-2}]$')
ax[1][1].set_xlabel('Probability\n distribution[-]')
ax[1][1].yaxis.set_label_coords(1.2, 0.5)

ax[1][1].yaxis.tick_right()
ax[1][1].yaxis.set_label_position("right")

# plot 2D histogram

ax[1][0].hist2d(a, b,bins=500,cmap = plt.cm.jet, norm=LogNorm())



ax[1][0].set_xlabel(r'$\alpha \;[Gy^{-1}]$')
ax[1][0].set_ylabel(r'$\beta \;[Gy^{-2}]$'+'\n')

#ax[1][0].set_xlim(0.2,0.4)
#ax[1][0].set_ylim(0.,0.05)



# common settings
for axx in ax.flatten():
    axx.grid()
    axx.minorticks_on()
    axx.grid(which='major', linestyle=':', linewidth='0.2', color='k')
    axx.grid(which='minor', linestyle=':', linewidth='0.2', color='k')

ax[0][1].set_visible(False)

ax[1][0].legend( ['correlation -0.936'],
          loc='upper left', frameon=False)


tmp=pd.DataFrame(list(zip(a,b)),
                          columns=['alpha','beta'])
tmp.to_csv("results/figure_5_2.csv")
fig.savefig(fname="results/figure_5_2.png",dpi= 700)

# Correlation

x=list()
y=list()
i=0  

fig,ax = plt.subplots(figsize = (10,6))

for k,v in d.items():

    for m,n in v.items():
        if m>0:
            #print(i,k,m)
            y.append(d[k][m]['alpha_dep'][i] /  d[k][0]['ref_alpha_dep'][i]-1)
            x.append(d[k][m]['LET']*d[k][0]['ref_beta_dep'][i]/d[k][0]['ref_alpha_dep'][i])
    
ax.plot(x,y, ".",c='black',markersize=15)

ax.set_xlabel('\n'+r'$LET \cdot \beta_{ref}\cdot \alpha^{-1}_{ref} \;[Gy\cdot  keV  \cdot\mu m^{-1}]$')
ax.set_ylabel(r'$(\alpha \cdot \alpha^{-1}_{ref})-1 \;[-]$'+'\n')

ax.grid()
ax.minorticks_on()
ax.grid(which='minor', linestyle=':', linewidth='0.2', color='k')


tmp=pd.DataFrame(list(zip(x,y)),
                          columns=['data_x','data_y'])
tmp.to_csv("results/figure_5_4.csv")

fig.savefig(fname="results/figure_5_4.png",dpi= 700)
